requirements.txt
streamlit
pandas
plotly
fastapi
uvicorn
PyJWT
python-multipart
opencv-python-headless
reportlab
pyPDF2
google-cloud-storage